Fix serialization/deserialization of externally tagged enums (see #1 for more info)
The serialization/deserialization optimization for single-element
tuple enum variants has turned out to be too aggressive and as a
result of this, incorrect output was being produced for enum
variants containing Vec (or other sequences).

This commit is a breaking change because it changes serialization
output in some cases and some inputs that used to be considered
as correct are no longer valid inputs. However, the breaking change
shouldn't be an issue in practice because these inputs/outputs were
most likely causing runtime errors already.